refactor: extract pack resolve() to shared app/resolver.rs
Both input and output usecases duplicated the resolve-by-id-or-name
logic. Extract to a single function in app/resolver.rs to eliminate
the duplication and make the logic easier to test and modify.

Closes #19
