+++ src/app/input_usecases.rs
@@ -1,7 +1,10 @@
 use std::sync::Arc;
 
 use crate::{
-    app::ports::{CodeExcerptPort, ListFilter, PackRepositoryPort},
+    app::{
+        ports::{CodeExcerptPort, ListFilter, PackRepositoryPort},
+        resolver::resolve_pack,
+    },
     domain::{
         errors::{DomainError, Result},
         models::{Pack, RefSpec},
@@ -49,32 +52,7 @@ impl InputUseCases {
     // ── identity resolution ───────────────────────────────────────────────────
 
     async fn resolve(&self, identifier: &str) -> Result<Pack> {
-        let identifier = identifier.trim();
-        if identifier.is_empty() {
-            return Err(DomainError::InvalidData(
-                "identifier (id or name) is required".into(),
-            ));
-        }
-        // Try by id first
-        if let Ok(id) = PackId::parse(identifier) {
-            if let Some(pack) = self.repo.get_by_id(&id).await? {
-                return Ok(pack);
-            }
-            return Err(DomainError::NotFound(format!(
-                "pack '{}' not found",
-                identifier
-            )));
-        }
-
-        // Fall back to name lookup
-        let name = PackName::new(identifier)?;
-        if let Some(pack) = self.repo.get_by_name(&name).await? {
-            return Ok(pack);
-        }
-        Err(DomainError::NotFound(format!(
-            "pack '{}' not found",
-            identifier
-        )))
+        resolve_pack(self.repo.as_ref(), identifier).await
     }
 
     async fn resolve_for_update(&self, identifier: &str, expected_revision: u64) -> Result<Pack> {
