+++ src/app/mod.rs
@@ -1,3 +1,4 @@
 pub mod input_usecases;
 pub mod output_usecases;
 pub mod ports;
+pub mod resolver;
